+++ CHANGELOG.md
@@ -3,12 +3,16 @@
 Thanks [@Zacam](https://github.com/Zacam) for the help with getting Chat Buttons Begone to the BD Store!
 
 ### Added
-- Option to remove the Game Collection / Wishlist in the profile popout [Profile Customizations]
-- Option to remove "new" indicator in the server rail [Servers]
+- Option to remove the Game Collection in the profile popout [Profile Customizations]
+- Option to remove the Wishlist in the profile popout [Profile Customizations]
+- Option to remove "new" indicator in the server rail [Servers and Channels]
 - Option to remove avatars from the server voice chat list [Voice]
+- Option to remove "blocked message(s)" in chat [Miscellaneous]
 
 ### Changed
 - Converted CSS selectors to use `BdApi.Webpack` to reduce the performance impact of CBB
+- Renamed "Servers" category to "Servers and Channels"
+    - Reorganized settings
 
 ### Fixed
 - Crash when launching the plugin with no config file
@@ -19,6 +23,8 @@ Thanks [@Zacam](https://github.com/Zacam) for the help with getting Chat Buttons
     - You can no longer switch between update branches. desktop-land-and-learn (/b) is now deprecated.
 - Option to enable Single Click File Select
     - Please use [SingleClickFileSelect](https://github.com/LancersBucket/SingleClickFileSelect) if you need this functionality
+- Option to remove Super Reaction button
+    - Button was moved from [Message Actions] into the emoji picker panel, so it didn't fit anymore
 
 ## v3.2.2 - December 25th, 2025
 ### Fixed